requirements.txt
nicegui~=1.4.5
pydantic~=2.5.2
loguru~=0.7.2
fastapi~=0.104.1
validators~=0.22.0
requests~=2.31.0
newspaper3k~=0.2.8
nltk~=3.8.1
openai~=1.3.9
readabilipy~=0.2.0
readability~=0.3.1
num2words~=0.5.13
lingua-language-detector~=2.0.2